+++ CHANGES.rst
@@ -1,3 +1,13 @@
+..
+    Copyright 2022 Rafe Kaplan
+    SPDX-License-Identifier: Apache-2.0
+
+    Updated: 2022-06-06
+
+Next Version
+============
+- Integration with readthe docs at https://conhead.readthedocs.io.
+
 v0.4.0
 ======
 - Provide alternate location for configuration file.